% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISDSdatasets-package.R
\docType{package}
\name{ISDSdatasets-package}
\alias{ISDSdatasets}
\alias{ISDSdatasets-package}
\title{ISDSdatasets: Datasets to supplement the Introduction to Statistics and Data Science textbook}
\description{
This package contains various datasets used for practice problems and examples in the Introduction to Statistics and Data Science textbook.
}
\author{
\strong{Maintainer}: Danielle Sass \email{danielle.sass@northwestern.edu} (\href{https://orcid.org/0000-0001-9755-738X}{ORCID})

}
\keyword{internal}
