% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba_sample}
\alias{nba_sample}
\title{A sample of career basketball game data for Michael Jordan, Kobe Bryant, and LeBron James}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 32 columns.
}
\usage{
nba_sample
}
\description{
This dataset contains a random stratified sample of 100 games of each player
in the nba dataset.
}
\keyword{datasets}
